Generalize DSpark across target architectures
