+++ mlx_vlm/speculative/drafters/dspark/__init__.py
@@ -1,0 +1,12 @@
+from .config import DSparkConfig as ModelConfig
+from .dspark import DSparkDraftModel
+from .dspark import DSparkDraftModel as Model
+from .dspark import VanillaMarkov, validate_dspark_target
+
+__all__ = [
+    "DSparkDraftModel",
+    "Model",
+    "ModelConfig",
+    "VanillaMarkov",
+    "validate_dspark_target",
+]
